added the basic framework for a render context

+++ Engine/Source/Application/Application.cpp
@@ -33,7 +33,7 @@ namespace MARS
 	{
 		while (bRunning)
 		{
-			glClearColor(0.1, 0.1, 0.1, 1);
+			glClearColor((GLfloat)0.1, (GLfloat)0.1, (GLfloat)0.1, 1.0);
 			glClear(GL_COLOR_BUFFER_BIT);
 
 			ImGuiLayerPtr->OnBegin();
